Updated module_sps_tx to free dynamically allocated memory
Uses the memory management queue to get the data for transmission and free it after a successful transmission.

The memory management queue contains the number of transmission attempts, so we don't need to declare that counter outside of the transmit loops.

since we're using pointers rather than the structs themselves, use arrow notation -> rather than .

Get rid of the DATA_partial_clear stuff and use the new dynamic memory implementation to free the memory.