Fixing the add_GPS crash
Instead of putting data into the transmission queue and setting the global pointers to NULL at the end of module_sps_park and module_sps_profile, we use the global pointers to add GPS data and move the data to the transmission queue at the end of module_sps_move_to_surface. This keeps data in the global pointers up until the transmission state, so that every function that needs to modify the data can access it.